fix: model type modified from int to long

+++ TestApp/MovieApi/Models/TheMoviedb/Movies/MovieInfo/Response/GetMovieInfoResponseModel.cs
@@ -23,7 +23,7 @@ namespace MovieApi.Models.TheMoviedb.Movies.MovieInfo.Response
         public List<ProductionCompany> production_companies { get; set; }
         public List<ProductionCountry> production_countries { get; set; }
         public string release_date { get; set; }
-        public int revenue { get; set; }
+        public long revenue { get; set; }
         public int runtime { get; set; }
         public List<SpokenLanguage> spoken_languages { get; set; }
         public string status { get; set; }
